refactor: 팀 체크리스트 뷰홀더 분리
